[Part] Move VectorAdapter into own file in PartGui
