+++ src/Mod/Part/App/CMakeLists.txt
@@ -216,6 +216,8 @@ SET(Features_SRCS
     AttachExtension.cpp
     PrismExtension.cpp
     PrismExtension.h
+    VectorAdapter.cpp
+    VectorAdapter.h
 )
 SOURCE_GROUP("Features" FILES ${Features_SRCS})
